+++ src/main/java/kr/hhplus/be/server/product/domain/entity/Product.java
@@ -20,10 +20,10 @@ public class Product {
     private String name;
 
     @Column(name = "price", nullable = false)
-    private Long price;
+    private Integer price;
 
     @Column(name = "stock", nullable = false)
-    private Long stock;
+    private Integer stock;
 
     public void decreaseStock(Integer quantity) {
         if(this.stock < quantity) {
@@ -32,7 +32,7 @@ public class Product {
         this.stock -= quantity;
     }
 
-    public void increaseStock(Long quantity) {
+    public void increaseStock(Integer quantity) {
         this.stock += quantity;
     }
 
@@ -40,11 +40,11 @@ public class Product {
         return this.stock > 0;
     }
 
-    public Long totalPrice(Integer quantity) {
+    public Integer totalPrice(Integer quantity) {
         return this.price * quantity;
     }
 
-    public static Product of(String name, Long price, Long stock) {
+    public static Product of(String name, Integer price, Integer stock) {
         return Product.builder()
                 .name(name)
                 .price(price)
@@ -52,14 +52,14 @@ public class Product {
                 .build();
     }
 
-    public static Product of(Long id, Long stock) {
+    public static Product of(Long id, Integer stock) {
         return Product.builder()
                 .id(id)
                 .stock(stock)
                 .build();
     }
 
-    public static Product of(Long id, String name, Long price, Long stock) {
+    public static Product of(Long id, String name, Integer price, Integer stock) {
         return new Product(id, name, price, stock);
     }
